CRITICAL FIX: Pawn promotion bug - Fix data field size in FastMoveGenerator::Move
- Changed Move::data from uint16_t to uint32_t to accommodate all move fields
- Move constructor was storing promotion data beyond 16-bit limit (bits 16-19)
- FastMoveGenerator now correctly generates and encodes promotion moves
- All promotion moves (queen, rook, bishop, knight) properly handled
- Perfect perft accuracy maintained for both starting position and complex positions
- Added comprehensive test cases for promotion move generation and validation
- Resolves 176-node discrepancy in complex positions with pawn promotions

Key issue: uint16_t data field could only store 16 bits but move encoding needed:
- from: bits 0-5, to: bits 6-11, piece: bits 12-15, promotion: bits 16-19
- promotion field was being truncated, resulting in promotion=0 for all moves

Performance: Still achieving ~7M NPS with perfect chess rule compliance

+++ src/FastMoveGenerator.h
@@ -7,7 +7,7 @@ class FastMoveGenerator {
 public:
     // Use integer move encoding like BBC instead of strings
     struct Move {
-        uint16_t data;
+        uint32_t data; // Changed from uint16_t to uint32_t to accommodate all fields
         
         Move() : data(0) {} // Default constructor
